wire AsyncReloadingClassificationEngine in blueprint.xml
